+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning].
 ## [Unreleased]
 - added new model and texture for all processing machines and the Requester
 - added compat recipe for MEGACells ([#21])
+- added info panels for the Universal Press and the Singularity
+  - brief explanation how to obtain them
+  - can be disabled in the config
 - improved texture of the Universal Press
 - improved texture of the Resonating Seed
 - Requester now requires a channel